resolves conflict for moved getAggregatedAvailability file
